fix: 🐛 do not synthesize/throw error on left branch map
Previously, the middleware map would send an error status and then call
next() with an error. This is problematic because the error handler
conventionally assumes that it needs to perform response actions,
leading to the header being sent multiple times. This is a semantic
change and therefore considered a breaking change. The left value that
is an [response, nextError] tuple is also no longer supported for the
same reason.

BREAKING CHANGE: 🧨 Map's left value `[response, nextError]` is no longer supported.